flashcards and oauths in new name space
