+++ app/controllers/home/oauths_controller.rb
@@ -1,0 +1,31 @@
+module Home
+  class OauthsController < ApplicationController
+    skip_before_action :require_login
+
+    def oauth
+      login_at(auth_params[:provider])
+    end
+
+    def callback
+      provider = auth_params[:provider]
+      if @user = login_from(provider)
+        redirect_to root_path, notice: t('.login', provider: provider.titleize)
+      else
+        begin
+          @user = create_from(provider)
+          reset_session
+          auto_login(@user)
+          redirect_to root_path, notice: t('.login', provider: provider.titleize)
+        rescue
+          redirect_to root_path, alert: t('.fail', provider: provider.titleize)
+        end
+      end
+    end
+
+    private
+
+    def auth_params
+      params.permit(:code, :provider)
+    end
+  end
+end
