+++ app/views/dashboard/flashcards/compare.js.erb
