+++ app/views/dashboard/flashcards/index.html.erb
